src/golang.org/x/net/html/testdata/webkit/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<q><div><style></style><address><q>
#errors
#document
| <html>
|   <head>
|   <body>
|     <q>
|     <div>
|       <q>
|         <style>
|       <address>
|         <q>
|         <q>
